T: merge `RsIncludeMacroResolveTest` into `CargoGeneratedItemsResolveTest`
