+++ src/test/kotlin/org/rustSlowTests/lang/resolve/CargoGeneratedItemsResolveTest.kt
@@ -7,14 +7,14 @@ package org.rustSlowTests.lang.resolve
 
 import com.intellij.openapi.util.io.FileUtil
 import com.intellij.testFramework.fixtures.impl.TempDirTestFixtureImpl
-import org.rust.cargo.RsWithToolchainTestBase
+import org.rust.MinRustcVersion
 import org.rust.fileTree
 import org.rust.ide.experiments.RsExperiments
 import org.rust.lang.core.psi.RsPath
 import org.rust.openapiext.runWithEnabledFeature
+import org.rustSlowTests.cargo.runconfig.RunConfigurationTestBase
 
-// TODO: merge with [org.rustSlowTests.cargo.runconfig.RsIncludeMacroResolveTest]
-class CargoGeneratedItemsResolveTest : RsWithToolchainTestBase() {
+class CargoGeneratedItemsResolveTest : RunConfigurationTestBase() {
 
     private val tempDirFixture = TempDirTestFixtureImpl()
 
@@ -28,7 +28,206 @@ class CargoGeneratedItemsResolveTest : RsWithToolchainTestBase() {
         super.tearDown()
     }
 
-    fun `test generated cfg option`() = runWithEnabledFeature(RsExperiments.EVALUATE_BUILD_SCRIPTS) {
+    @MinRustcVersion("1.32.0")
+    fun `test include in workspace project`() = withEnabledFetchOutDirFeature {
+        val testProject = buildProject {
+            toml("Cargo.toml", """
+                [package]
+                name = "intellij-rust-test"
+                version = "0.1.0"
+                authors = []
+            """)
+            rust("build.rs", """
+                use std::env;
+                use std::fs::File;
+                use std::io::Write;
+                use std::path::Path;
+
+                fn main() {
+                    let out_dir = env::var("OUT_DIR").unwrap();
+                    let dest_path = Path::new(&out_dir).join("hello.rs");
+                    let mut f = File::create(&dest_path).unwrap();
+
+                    f.write_all(b"
+                        pub fn message() -> &'static str {
+                            \"Hello, World!\"
+                        }",
+                    ).unwrap();
+                }
+            """)
+            dir("src") {
+                rust("main.rs", """
+                    include!(concat!(env!("OUT_DIR"), "/hello.rs"));
+
+                    fn main() {
+                        println!("{}", message());
+                                        //^
+                    }
+                """)
+            }
+        }
+        buildProject()
+
+        runWithInvocationEventsDispatching("Failed to resolve the reference") {
+            testProject.findElementInFile<RsPath>("src/main.rs").reference.resolve() != null
+        }
+    }
+
+    // https://github.com/intellij-rust/intellij-rust/issues/4579
+    @MinRustcVersion("1.32.0")
+    fun `test do not overflow stack 1`() = withEnabledFetchOutDirFeature {
+        val testProject = buildProject {
+            toml("Cargo.toml", """
+                [package]
+                name = "intellij-rust-test"
+                version = "0.1.0"
+                authors = []
+            """)
+            rust("build.rs", """
+                use std::env;
+                use std::fs::File;
+                use std::io::Write;
+                use std::path::Path;
+
+                fn main() {
+                    let out_dir = env::var("OUT_DIR").unwrap();
+                    let dest_path = Path::new(&out_dir).join("main.rs");
+                    let mut f = File::create(&dest_path).unwrap();
+
+                    f.write_all(b"
+                        pub fn message() -> &'static str {
+                            \"Hello, World!\"
+                        }",
+                    ).unwrap();
+                }
+            """)
+            dir("src") {
+                rust("main.rs", """
+                    include!(concat!(env!("OUT_DIR"), "/main.rs"));
+
+                    fn main() {
+                        println!("{}", message());
+                                        //^
+                    }
+                """)
+            }
+        }
+        buildProject()
+
+        runWithInvocationEventsDispatching("Failed to resolve the reference") {
+            testProject.findElementInFile<RsPath>("src/main.rs").reference.resolve() != null
+        }
+    }
+
+    // https://github.com/intellij-rust/intellij-rust/issues/4579
+    @MinRustcVersion("1.32.0")
+    fun `test do not overflow stack 2`() = withEnabledFetchOutDirFeature {
+        val testProject = buildProject {
+            toml("Cargo.toml", """
+                [workspace]
+                members = [
+                    "intellij-rust-test-1",
+                    "intellij-rust-test-2"
+                ]
+            """)
+            dir("intellij-rust-test-1") {
+                toml("Cargo.toml", """
+                    [package]
+                    name = "intellij-rust-test-1"
+                    version = "0.1.0"
+                    authors = []
+
+                """)
+                rust("build.rs", """
+                    use std::env;
+                    use std::fs::File;
+                    use std::io::Write;
+                    use std::path::Path;
+
+                    fn main() {
+                        let out_dir = env::var("OUT_DIR").unwrap();
+                        let dest_path = Path::new(&out_dir).join("lib.rs");
+                        let mut f = File::create(&dest_path).unwrap();
+
+                        f.write_all(b"
+                            pub fn message() -> &'static str {
+                                \"Hello, World!\"
+                            }",
+                        ).unwrap();
+                    }
+                """)
+                dir("src") {
+                    rust("lib.rs", """
+                        include!(concat!(env!("OUT_DIR"), "/lib.rs"));
+
+                        fn main() {
+                            println!("{}", message());
+                                            //^
+                        }
+                    """)
+                }
+            }
+            dir("intellij-rust-test-2") {
+                toml("Cargo.toml", """
+                    [package]
+                    name = "intellij-rust-test-2"
+                    version = "0.1.0"
+                    authors = []
+                """)
+                rust("build.rs", """
+                    use std::env;
+                    use std::fs::File;
+                    use std::path::Path;
+
+                    fn main() {
+                        let out_dir = env::var("OUT_DIR").unwrap();
+                        let dest_path = Path::new(&out_dir).join("lib.rs");
+                        let mut f = File::create(&dest_path).unwrap();
+                    }
+                """)
+                dir("src") {
+                    rust("lib.rs", """
+                        include!(concat!(env!("OUT_DIR"), "/lib.rs"));
+                    """)
+                }
+            }
+        }
+        buildProject()
+
+        runWithInvocationEventsDispatching("Failed to resolve the reference") {
+            testProject.findElementInFile<RsPath>("intellij-rust-test-1/src/lib.rs").reference.resolve() != null
+        }
+    }
+
+    @MinRustcVersion("1.32.0")
+    fun `test include in dependency`() = withEnabledFetchOutDirFeature {
+        val testProject = buildProject {
+            toml("Cargo.toml", """
+                [package]
+                name = "intellij-rust-test"
+                version = "0.1.0"
+                authors = []
+
+                [dependencies]
+                code-generation-example = "0.1.0"
+            """)
+            dir("src") {
+                rust("lib.rs", """
+                    fn main() {
+                        println!("{}", code_generation_example::message());
+                                                               //^
+                    }
+                """)
+            }
+        }
+        buildProject()
+        runWithInvocationEventsDispatching("Failed to resolve the reference") {
+            testProject.findElementInFile<RsPath>("src/lib.rs").reference.resolve() != null
+        }
+    }
+
+
+    fun `test generated cfg option`() = withEnabledEvaluateBuildScriptsFeature {
         val libraryDir = tempDirFixture.getFile(".")!!
         val library = fileTree {
             toml("Cargo.toml", """
@@ -94,7 +293,7 @@ class CargoGeneratedItemsResolveTest : RsWithToolchainTestBase() {
         }.checkReferenceIsResolved<RsPath>("src/main.rs", toFile = ".../src/enabled.rs")
     }
 
-    fun `test generated environment variables`() = runWithEnabledFeature(RsExperiments.EVALUATE_BUILD_SCRIPTS) {
+    fun `test generated environment variables`() = withEnabledEvaluateBuildScriptsFeature {
         buildProject {
             toml("Cargo.toml", """
                 [package]
@@ -129,4 +328,10 @@ class CargoGeneratedItemsResolveTest : RsWithToolchainTestBase() {
             """)
         }.checkReferenceIsResolved<RsPath>("src/main.rs", toFile = ".../foo/bar/hello.rs")
     }
+
+    private fun withEnabledFetchOutDirFeature(action: () -> Unit) =
+        runWithEnabledFeature(RsExperiments.FETCH_OUT_DIR, action)
+
+    private fun withEnabledEvaluateBuildScriptsFeature(action: () -> Unit) =
+        runWithEnabledFeature(RsExperiments.EVALUATE_BUILD_SCRIPTS, action)
 }
